move elements into scene
