+++ src/ts/app/view/viewport/scene/elements/base.ts
@@ -1,4 +1,4 @@
-export interface ViewportElementCallbacks {
+export type ViewportElementCallbacks = {
     onCreate: (symbol: THREE.Object3D) => void
     onDispose: (symbol: THREE.Object3D) => void
 }
