+++ src/ts/app/view/viewport/scene/elements/connection.ts
@@ -1,5 +1,5 @@
 import * as THREE from 'three'
-import { Landmark } from '../base'
+import { Landmark } from '../../base'
 import { ViewportElementCallbacks } from './base'
 
 const LINE_COLOR = 0xffff00
